rewrite(web): slice 9s.2 — WebSocket platform primitive
This slice ships the platform-level WS connection that future
tile slices (terminal, Claude feed, third-party tiles) all
ride on. Per the user's "treat the terminal as just another
tile" framing: the WS connection is a platform service, not
a terminal feature. Anything the terminal will need from this
connection is exactly what every other tile gets — there's no
privileged code path.

The server side is essentially complete from earlier slices
(9c shipped the transport abstraction + WS impl + typed
protocol; 9d moved to CBOR; 9e added the handshake state
machine + session attach/output/reconnect). What this slice
adds is the WASM client + the move of the protocol types to
shared so the client can speak to it.

Wire types — moved to `katulong_shared::wire`
- `ClientMessage` (Ping, HelloAck, Attach, Input, Resize) and
  `ServerMessage` (Hello, Pong, Attached, Output, Error,
  OutputGap) + `PROTOCOL_VERSION` were defined in
  `crates/server/src/transport/message.rs`. Moved them into
  `crates/shared/src/wire.rs` alongside the auth + tile types
  so the WASM crate can import the same definitions the
  server publishes — no hand-rolled mirror, no drift risk.
- `crates/server/src/transport/message.rs` becomes a re-
  export shim. Server-internal call sites
  (`use crate::transport::message::*`) keep working with no
  changes; the protocol-level documentation moves with the
  types.
- 19 round-trip + serde-shape tests moved from the inline
  `#[cfg(test)] mod tests` in transport/message.rs to a new
  integration-test file `crates/shared/tests/session_message.rs`.
  Test count and coverage are preserved (server unit count
  drops from 188 to 169; shared tests gain 19; total
  unchanged).
- `serde_bytes` dep added to katulong-shared (required by the
  `#[serde(with = "serde_bytes")]` annotation on
  Input.data / Output.data); `ciborium` added to
  katulong-shared's dev-deps for the CBOR round-trip tests.

WASM-side platform primitive — `crates/web/src/ws.rs`
- `ws::install(auth_state, connection_status)` called from
  `<App/>` after both contexts are provided. Uses
  `create_effect` to watch `AuthPhase`; on the first
  `SignedIn` transition, spawns the connection lifecycle.
- Connection lifecycle:
  1. Resolve same-origin WS URL (`ws://` or `wss://` based
     on page protocol).
  2. Open via `gloo_net::websocket::futures::WebSocket`.
  3. Receive Hello, validate `protocol_version`, send
     HelloAck.
  4. `set_connected.set(true)` — handshake complete.
  5. Drain the inbound stream until it ends; on close,
     `set_connected.set(false)`.
- The inbound drain currently logs unknown messages and
  discards. The tile-side subscribe API (where TerminalTile
  filters for `Output { session_id }` messages, future
  ClaudeFeed filters for topic events, etc.) is the 9s.3
  scope — designing it against zero consumers would be the
  framework-not-protocol failure mode the
  `feedback_no_premature_generalization` memory warns
  against.
- All errors log via `web_sys::console::warn_1` so dev triage
  is observable in browser devtools without surfacing
  internal state into user-visible UI.

`ConnectionStatus` shape — both halves exposed
- Was `pub struct ConnectionStatus(ReadSignal<bool>)` with
  `_set_connected` discarded at App-root. The PR #681
  architecture nit flagged this as a deferred shape gap.
  Now `pub struct ConnectionStatus { connected, set_connected }`
  matching `AuthState`'s named-field shape.
- The WS module owns the writer; the Header and StatusTile
  destructure with `..` so they only see the reader.

Bundle weight
- 591 KB → 752 KB (+161 KB). Sources: `gloo-net`'s
  `websocket` feature pulls in `wasm-streams` and the
  futures Stream/Sink wrappers; `ciborium` adds the CBOR
  codec; `futures-util` for `.next().await` on the
  WebSocket stream. The platform primitive is the marginal
  cost; future tiles consume it without re-paying.

E2E coverage
- New assertion in `webauthn.e2e.js` test 3 (session restore):
  after the post-auth view renders, the header's status
  indicator must transition to `data-status="connected"`
  within 5s. This catches a regression in the WS lifecycle
  without requiring any tile to be focused — the indicator
  is one of two `ConnectionStatus` consumers (StatusTile is
  the other) and it's always visible.
- Existing 13 e2e tests pass unchanged. Auth-only smoke /
  rejection tests stub `/api/auth/status` to look
  unauthenticated, so they never reach `SignedIn` and the
  WS doesn't connect — the indicator stays "disconnected"
  as before.

What does NOT land
- Tile-side send/subscribe API. Lands in 9s.3 with the real
  terminal — which is the first concrete consumer.
- Reconnect on disconnect. Today the connection is open-
  once-per-page-load. A future slice adds backoff +
  reconnect.
- Logout-side disconnect. Tied to the future logout slice
  (which already has the layout-clear obligation flagged).

Verified
- `cargo build --release` clean.
- `cargo test --release` green: 297 tests across 11
  binaries (was 297; the message tests moved within shared
  but the total is unchanged).
- `trunk build --release` clean.
- Full Rust e2e (13 tests) green on a fresh data dir AND on
  a re-run with dirty data dir. The new WS-handshake
  assertion in test 3 fires correctly — handshake completes
  before the timeout.

Slice scope
9s.2 is the platform-vs-tile-feature commitment. The next
slice (9s.3, real terminal) is the first consumer; the
shape it consumes (the to-be-designed send/subscribe API)
is what tells us whether the platform shape is right.
