+++ crates/shared/src/wire.rs
@@ -334,3 +334,200 @@ impl TileLayout {
         }
     }
 }
+
+// =====================================================================
+// Session protocol — typed wire for the WebSocket session layer.
+//
+// Every message crossing the transport boundary is one of these
+// variants — no `serde_json::Value` catch-alls, no loosely-typed
+// payloads. The Node scar that motivated this (`9dc7c78`) was a
+// terminal that accepted unvalidated JSON and passed fields straight
+// to PTY resize/input — `"999999"` where a number was expected could
+// cause type confusion. Rust's types + serde's strictness close that
+// class of bug at parse time; we just have to be disciplined about
+// not reintroducing `Value` as an escape hatch.
+//
+// **Wire format: CBOR.** Messages encode via CBOR (`ciborium`) into
+// binary frames. JSON is used only for the HTTP API (cookies, error
+// envelopes, token CRUD). Uniform binary over both WS and (future)
+// WebRTC DataChannel; no base64 overhead for byte payloads; smaller
+// wire; same serde derives.
+//
+// **Byte-string encoding for `Input`/`Output`.** `Input { data }` and
+// `Output { data }` carry raw terminal bytes. Their `data` fields are
+// annotated `#[serde(with = "serde_bytes")]` so CBOR emits a major-
+// type-2 byte string instead of the serde-default array of small
+// integers. Without the annotation, a 1 KiB paste would encode as
+// ~1–2 KiB of integer elements; with it the payload rides the wire
+// literally.
+//
+// **Strictness flags.**
+// - `#[serde(tag = "type", rename_all = "snake_case")]` — every
+//   message carries a `type` discriminator and snake_case field
+//   names. Missing discriminator fails parsing.
+// - `#[serde(deny_unknown_fields)]` on `ClientMessage` only — inbound
+//   strictness is a security property; outbound `ServerMessage` stays
+//   lenient so older Rust consumers (tests, federation relays) can
+//   deserialize newer-server output without hard-failing on unknown
+//   fields.
+//
+// **Handshake.** Three-step gate before terminal I/O is valid:
+//   1. Server → Client: `Hello { protocol_version }` on upgrade.
+//   2. Client → Server: `HelloAck { protocol_version }`. Server re-
+//      validates; mismatch → `Error { code:
+//      "protocol_version_mismatch", ... }` and close.
+//   3. Client → Server: `Attach { session, cols, rows }`. Server
+//      replies `Attached`.
+// Only after `Attached` are `Input`/`Resize` accepted.
+// `Ping`/`Pong` is allowed in every phase (liveness is orthogonal).
+//
+// **Why these types live in `shared`.** Server (sends + receives) and
+// WASM client (sends + receives) both consume them. Originally lived
+// in `crates/server/src/transport/message.rs`; moved here so the
+// types are the single source of truth and a server schema change
+// can't drift from a hand-rolled WASM mirror.
+// =====================================================================
+
+/// Current protocol version. Clients check this against their
+/// expected value and refuse to proceed on mismatch. Bumped when a
+/// non-backwards-compatible change lands (new required field, removed
+/// variant, changed semantics).
+pub const PROTOCOL_VERSION: &str = "katulong/0.1";
+
+/// Messages sent by the client to the server.
+///
+/// Deserialized from each inbound CBOR binary frame. Text frames are
+/// rejected by the transport layer. Byte payloads (terminal input,
+/// future image paste) carry directly via CBOR's byte-string type —
+/// no base64 wrapper needed.
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(tag = "type", rename_all = "snake_case", deny_unknown_fields)]
+pub enum ClientMessage {
+    /// Client heartbeat. Server echoes with `Pong` carrying the same
+    /// `nonce` — lets the client measure round-trip latency without
+    /// assuming a specific transport's heartbeat primitive (WS Ping
+    /// frames vs WebRTC DC keep-alives differ).
+    Ping { nonce: u64 },
+
+    /// Acknowledge the server's `Hello` and confirm the client speaks
+    /// the same protocol version. Sent exactly once, as the first
+    /// message after the client receives `Hello`. The server re-
+    /// checks the version and closes on mismatch.
+    HelloAck { protocol_version: String },
+
+    /// Bind this transport to a tmux session. `session` is the tmux
+    /// session name (validated by `SessionManager`); `cols`/`rows`
+    /// are the client's current window dimensions (clamped
+    /// defensively server-side per `session::dims`). Sent exactly
+    /// once after `HelloAck`. The server creates the session if it
+    /// doesn't exist (or attaches to the existing one) and replies
+    /// with `Attached`.
+    ///
+    /// `resume_from_seq`: reconnect hint. Absent (or `None`) means
+    /// "fresh attach — I have no prior state." `Some(N)` means "I
+    /// last received bytes through seq N; please replay bytes
+    /// `(N..last_seq]` if you still have them." If the server's ring
+    /// has lost bytes below `N`, it emits `OutputGap` before the
+    /// replay so the client clears its terminal first. Omit on first
+    /// attach; echo the seq from `Attached.last_seq` on reconnect.
+    Attach {
+        session: String,
+        cols: u16,
+        rows: u16,
+        #[serde(default)]
+        resume_from_seq: Option<u64>,
+    },
+
+    /// Keystroke / paste input destined for the PTY. Only valid after
+    /// `Attached`.
+    Input {
+        #[serde(with = "serde_bytes")]
+        data: Vec<u8>,
+    },
+
+    /// Window-resize notification. Only valid after `Attached`.
+    /// Forwarded to the session manager, which clamps and issues a
+    /// tmux `refresh-client -C`. Do NOT send on every keystroke — see
+    /// `session::dims` for the SIGWINCH-storm history.
+    Resize { cols: u16, rows: u16 },
+}
+
+/// Messages sent by the server to the client.
+///
+/// Serialized as CBOR binary frames. Clients treat any message with
+/// an unknown `type` field as a forward-compat signal and log but
+/// don't reject — we want server → client additions to be deployable
+/// without client pinning.
+///
+/// **Asymmetry with `ClientMessage`.** `deny_unknown_fields` is
+/// deliberately OMITTED here. The strict boundary is INBOUND: the
+/// server rejects unknown client input because that's untrusted data.
+/// Outbound messages from the server are produced by our own code;
+/// relaxing this direction lets older Rust clients deserialize
+/// newer-server output without hard-failing on fields they don't
+/// understand.
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(tag = "type", rename_all = "snake_case")]
+pub enum ServerMessage {
+    /// Sent immediately after a successful transport upgrade. Tells
+    /// the client the connection is live and which protocol version
+    /// the server speaks. The client can refuse to proceed if the
+    /// version doesn't match its own expectation.
+    Hello { protocol_version: String },
+
+    /// Response to a client `Ping`. Echoes the same nonce.
+    Pong { nonce: u64 },
+
+    /// Confirms that this transport is bound to `session`, which has
+    /// been resized to the clamped `cols`/`rows`. The client should
+    /// use these (not what it requested) as the authoritative
+    /// dimensions for its local renderer.
+    ///
+    /// `last_seq` is the pane's current `total_written` — the byte
+    /// counter the ring is at when this `Attached` ships. Fresh-
+    /// attach client seeds its local seq from here. Reconnect client
+    /// uses it as the upper bound of the replay they're about to
+    /// receive. `0` on a fresh pane (no output ever produced).
+    Attached {
+        session: String,
+        cols: u16,
+        rows: u16,
+        last_seq: u64,
+    },
+
+    /// PTY output chunk. `seq` is the pane's `total_written`
+    /// (cumulative decoded byte offset) at the end of this chunk — a
+    /// per-PANE monotonic counter that continues across subscriber
+    /// displace/reconnect boundaries.
+    ///
+    /// **Seq contract.** `seq` is 1-based at the first-byte level:
+    /// the first outbound byte of a pane carries `end_seq = 1`; seq =
+    /// 0 is reserved to mean "no output produced yet" (used in
+    /// `Attached.last_seq` on a cold pane). Clients counting gaps see
+    /// monotonic progress across reconnect as long as the server
+    /// process has not restarted; a decrease means a fresh server
+    /// instance and the client must treat it as a hard reset.
+    Output {
+        #[serde(with = "serde_bytes")]
+        data: Vec<u8>,
+        seq: u64,
+    },
+
+    /// Protocol-level error. The server sends this right before
+    /// closing the transport so the client sees a concrete reason in
+    /// logs rather than an opaque WS close frame.
+    Error { code: String, message: String },
+
+    /// Reconnect replay had a gap: the client's `resume_from_seq` was
+    /// older than the oldest byte still in the pane's ring.
+    /// `available_from_seq` is the oldest byte we can replay (>
+    /// the client's request); `last_seq` matches `Attached.last_seq`
+    /// for convenience. The client MUST clear its terminal / restart
+    /// its renderer before applying any subsequent `Output` — cursor
+    /// positions and in-flight escape sequences from the lost window
+    /// can't be inferred.
+    OutputGap {
+        available_from_seq: u64,
+        last_seq: u64,
+    },
+}
